intl/l10n/test/dom/test_docl10n_ready_rejected.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test mozIDOMLocalization.ready rejected state</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <link rel="localization" href="path/to_non_existing.ftl"/>
  <script type="application/javascript">
  "use strict";
  SimpleTest.waitForExplicitFinish();

  document.addEventListener("DOMContentLoaded", async function() {
      document.l10n.ready.then(() => {
        is(1, 2, "the ready should not resolve");
        SimpleTest.finish();
      }, (err) => {
        is(1, 1, "the ready should reject");
        SimpleTest.finish();
      });
  });
  </script>
</head>
<body>
  <h1 data-l10n-id="non-existing-id"></h1>
</body>
</html>
